requirements.txt


# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aiofiles==24.1.0
jinja2==3.1.4

# ─── Configuration ────────────────────────────────────────────────────────────
python-dotenv==1.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0

# ─── Database (SQLite prototype → PostgreSQL production) ─────────────────────
sqlalchemy==2.0.32
# psycopg2-binary==2.9.9       #PostgreSQL

# ─── AI & OCR ─────────────────────────────────────────────────────────────────
google-genai>=1.0.0
pytesseract==0.3.10
Pillow==10.4.0

# ─── PDF Processing ───────────────────────────────────────────────────────────
reportlab==4.2.2
pdfplumber==0.11.4
pypdf==4.3.1
# pdf2image==1.17.0             # scanned PDF OCR (poppler)

# ─── Authentication & Security ────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# ─── Image Forensics ──────────────────────────────────────────────────────────
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
scipy==1.14.1

# ─── Utilities ────────────────────────────────────────────────────────────────
httpx==0.27.0
